refactor: change detail page from dialog to link
