+++ components/demands/demand-preview.tsx
@@ -1,10 +1,7 @@
 'use client';
 
-import { useState } from 'react';
-import { useRouter } from 'next/navigation';
 import { DemandPreviewFragment } from '@/generated/graphql';
-import DemandDetail from './demand-detail';
-import DealPreviewLayout from '../deals/deal-preview.layout';
+import DealPreview from '../deals/deal-preview';
 
 interface Props {
   demand: DemandPreviewFragment;
@@ -12,34 +9,15 @@ interface Props {
 }
 
 export default function DemandPreview({ demand, type }: Props) {
-  const router = useRouter();
-  const [open, setOpen] = useState(false);
-
-  const handleOpen = () => {
-    setOpen(!open);
-    window.history.pushState(
-      {},
-      ``,
-      `/user/${demand.buyer.username}/demands/${demand.slug}`,
-    );
-  };
-
-  const handleClose = () => {
-    setOpen(!open);
-    router.back();
-  };
-
   return (
-    <DealPreviewLayout
+    <DealPreview
+      deal="demands"
       type={type}
       name={demand.name}
       price={demand.price}
       createdAt={demand.createdAt}
-      open={open}
-      handleOpen={handleOpen}
-      handleClose={handleClose}
-    >
-      <DemandDetail slug={demand.slug!} />
-    </DealPreviewLayout>
+      username={demand.buyer.username}
+      slug={demand.slug!}
+    />
   );
 }
