+++ components/swaps/swap-preview-preview.tsx
@@ -1,11 +1,8 @@
 'use client';
 
-import { useState } from 'react';
-import { useRouter } from 'next/navigation';
 import { SwapPreviewFragment } from '@/generated/graphql';
-import SwapDetail from './swap-detail';
 import SwapName from './swap-name';
-import DealPreviewLayout from '../deals/deal-preview.layout';
+import DealPreview from '../deals/deal-preview';
 
 interface Props {
   swap: SwapPreviewFragment;
@@ -13,35 +10,16 @@ interface Props {
 }
 
 export default function SwapPreview({ swap, type }: Props) {
-  const router = useRouter();
-  const [open, setOpen] = useState(false);
-
-  const handleOpen = () => {
-    setOpen(!open);
-    window.history.pushState(
-      {},
-      ``,
-      `/user/${swap.proposer.username}/swaps/${swap.slug}`,
-    );
-  };
-
-  const handleClose = () => {
-    setOpen(!open);
-    router.back();
-  };
-
   return (
-    <DealPreviewLayout
+    <DealPreview
+      deal="swaps"
       type={type}
       thumbnail={swap.thumbnail}
       name={<SwapName name0={swap.name0} name1={swap.name1} />}
       price={swap.price}
       createdAt={swap.createdAt}
-      open={open}
-      handleOpen={handleOpen}
-      handleClose={handleClose}
-    >
-      <SwapDetail slug={swap.slug!} />
-    </DealPreviewLayout>
+      username={swap.proposer.username}
+      slug={swap.slug!}
+    />
   );
 }
